requirements.txt
google-cloud-logging
datetime
alive_progress
google-cloud-resource-manager 
tqdm
google-cloud-storage